In add_delete and add_rr, the name was not being written before the RRTYPE, etc., and so of course that didn't parse.  Updated to write the name first.
The add_rr function is allowed to take a nonexistent RR, which means don't write anything, but this test hadn't been done.   Updated to do the test.

In construct_update, the towire struct was incompletely initialized; updated to memset it to zero before filling in non-default fields.

In construct_update, update length was not being set; updated to set it.

Add some ifdef'd debug code in update_send to decode the update before sending it to make sure it's well-formed.

In update_connect_callback, update_length and update_max had been confused, and so the update was seen as having zero space and wasn't being constructed.   The length was also incorrect.   Fixed to use the correct length in the correct place.

In update_connect_callback, the update zone was not being retained, and so the update was being proposed to be applied to the root zone rather than the correct update zone.