Fix API to support port randomization changes in Core.
When source port randomization was added, mDNSPosix wasn't updated to work with
the new code--the APIs were just updated but never filled in.   As a result, DNS
resolution stopped working.   This commit does the minimum possible change to
make it work again, but does not actually add source port randomization.  That
will need to be added in a later commit.