Add support for specifying more than one IP address for listening.
This is required because DNS answers have to come from the IP address
to which the question was sent.

Fix some residual MY_NAME code that didn't get noticed earlier.

If we receive a query for my_name, don't recurse: answer it directly
even if it's not in an interface zone.  (Maybe shouldn't claim
authority, though).

Support listening on IPv6 as well as IPv4.

Add a --tls-fail flag that causes incoming TLS connections to fail after
the connection is accepted, for testing.   Not useful in production.