Fix the multi-record read logic for inotifyReadCallback().
Simply watching /etc/resolv.conf using inotify doesn't work because
inotify won't let us watch a file that isn't there.   In order to
make inotify detect all possible changes, we have to watch /etc for
changes, and if /etc/resolv.conf is a symlink, we need to watch it,
the file it points to, and the directory containing that file.

Additionally, if the directory containing either the symlink or the
file doesn't exist, we have to watch the closest parent that does
exist, which in the degenerate case is the root directory.