A previous change added the possibility to serve names out of zones for which the proxy isn't authoritative.   Examples of this would be the local browsing domain information and the hostname the NS and SOA records point to as the authoritative server for the zone, by default discoveryproxy.home.arpa., which is not in a zone operated by the server.
An effect of this was that all names in that zone would be intercepted.   This
prevented reverse lookups from working, for example, because the whole in-addr.arpa
zone was being intercepted.

Additionally, a change previously made to handle the case where the the nameserver
name was not in a zone for which the proxy was authoritative was done wrong, so that
the A record and AAAA record was being mangled.

This commit fixes it so that, first of all, the nameserver's addresses are published
correctly.   And secondly, intercepted zones for which the proxy is not authoritative
are now handled correctly and are not resolved using Multicast DNS.